+++ drel/core/utils.py
@@ -1,5 +1,6 @@
 from typing import Dict
 
+from django.http import HttpRequest
 from marshmallow import Schema, fields
 
 from drel.utils import datetime_to_week_range
@@ -16,3 +17,7 @@ class EmailUserSchema(Schema):
 def get_index_name() -> str:
     week_start, week_end = datetime_to_week_range()
     return f"logs-{week_start.date()}-{week_end.date()}"
+
+
+def log_only_post(request: HttpRequest) -> bool:
+    return request.method != "POST"