+++ tests/test_requests.py
@@ -5,6 +5,7 @@ import responses
 from requests import Request
 from requests import Session
 
+from drel.core.config import ELASTIC_SEARCH_RUN_TESTS
 from drel.core.es import get_from_es
 from drel.core.models import FullRequestLog
 from drel.core.models import RequestLog
@@ -72,7 +73,7 @@ def test_full_request_to_log(
 
 
 @pytest.mark.skipif(
-    not os.getenv("ELASTIC_SEARCH_RUN_TESTS"),
+    not ELASTIC_SEARCH_RUN_TESTS,
     reason="Set ELASTIC_SEARCH_RUN_TESTS env to enable Elastic Search tests",
 )
 def test_requests_log_insert_to_es(